Fix a couple harmless BCC warnings.

+++ src/syncterm/conn.c
@@ -22,7 +22,7 @@ int conn_recv(char *buffer, size_t buflen, unsigned timeout)
 	BYTE	*p;
 	BOOL	data_waiting;
 	static BYTE	*telnet_buf=NULL;
-	static int tbsize=0;
+	static size_t tbsize=0;
 
 	if(con_type == CONN_TYPE_TELNET) {
 		if(tbsize < buflen) {
